Add JSON file with national applicability
Adds a test content item in a JSON with national applicability added to
every entry in the timeline.

This file will be loaded locally instead of the content item to test the
nation toggling feature for the timeline works correctly on the
coronavirus landing page.

This file can be removed when the changes are ready to go live.