Add a new class that extends ContentItem
This class loads the content item from a json file rather than
content-store.

Uses inheritance so that we can add a new load method, but keep all of
the other properties of the ContentItem model, and so that the code is
easy to remove.

The landing page controller will decide which content item class to
call.