Load test content_item if "nation" param exists
Loads the test content item from the JSON file if a "nation" param is
passed into the querystring. It does not matter what value the param has
as long as it exists.

The view will use the existence of the "national_applicability" in the
timeline entries to determine whether or not to show the radio buttons,
not this check. That means that when this feature is live and the live
content item also has "national_applicability" the feature will
automatically start to work.

Adds a guard clause to make sure that the test data can only loaded in
development environments too to avoid embarrassment. This means the
feature can't be tested in Integration.

The call to CoronavirusTimelineNationsContentItem is being wrapped in a
method otherwise Brakeman to raises the following warning:

== Warning Types ==

Cross-Site Scripting: 1

== Warnings ==

Confidence: Weak
Category: Cross-Site Scripting
Check: CrossSiteScripting
Message: Unescaped model attribute
Code: JSON.pretty_generate(presenter.faq_schema((CoronavirusTimelineNationsContentItem.load.to_hash or content_item.to_hash)))
File: app/views/coronavirus_landing_page/components/shared/_schema.html.erb
Line: 2

[1]: https://brakemanscanner.org/docs/warning_types/cross_site_scripting/