Move timeline_nation_items method to helper
This method is only used in the view, so it seems like a good candidate to be moved into
the view helper. Moving the method also makes it easier to add tests for it.

timeline_nation_items relied on knowing the value of the `timeline_nation` param. This is
now being stored in an instance variable so that it can be accessed more easily