Add method to check for national applicability
This method will be used by the views to determine whether to show the
new timeline for UK nations and the radio button picker, or the existing
full timeline.

This check is needed until the content_item is updated to have the
national applicability for each timeline entry, after which point it can
be removed.