Get tags for all nations timeline entry applies to
The design asked that if a timeline entry applies to all of the UK
nations, that "UK Wide" be returned instead.

`timeline_nation_tags` will be returned as part of `timeline_for_nation`
in later commits so that it can be rendered more easily in the view.

Uses `html_safe` to stop the html content in the nation tags from being
escaped when the page is rendered.