Use presenter to return the timelines for each nation as one object
This reduces the amount of logic in the view and hopefully makes it easier to read.
The test is very explicit as I think it helps to understand what is being returned if you
can see it in its entirety.

The existing timeline_for_nation method has been made private, so it's tests have been
removed.