Move rest of timeline content into locale file
Also moves the html to render the links for the additional country
guidance into the locale file to make the view a bit tidier.

We're importing `ActionView::Helpers::SanitizeHelper` in strange way in
the feature tests so that the class methods [1] will be added to the
module rather than the class that is including the module.

Otherwise we get the following error on `strip_tags`:

```
NoMethodError:
       undefined method `full_sanitizer' for #<Class:0x00007f9242062508>
```

The other fix would have been to do the include in the test class rather
than the helper module, which would have been odd as the method is being
used in the helper.

[1]: https://github.com/rails/rails/blob/398561bfbb7d156a141918df20b2fc622a94975d/actionview/lib/action_view/helpers/sanitize_helper.rb#L124-L169