Move `timeline_nations_items` method to presenter
We'll need to add more methods to the presenter in later commits to
display the COVID-19 timeline for UK nations, so it makes sense to move
the presentation of the radio buttons here as well.

The order of the countries has been changed to be alphabetical to
prevent any controversy, and the view for timeline has been updated to
accept the entire presenter object so that it can access all presenter
methods.