Apply JS and hide radio buttons under details
- make the radio buttons for nation picker on the covid landing page
work using JS, if available
- dynamically updates the page based on immediate radio choice

The tracking has been added as part of the details component so it no
longer needs to be directly applied to the form.

The "Showing updates..." text needs a guard as `params[:nation]` may not
exist in the query string, and `titleize` cannot be called on a null
object. This will be fixed in a later commit when the default nation
will be set to England.

Co-authored-by: Andy Sellick <andy.sellick@digital.cabinet-office.gov.uk>