Show timeline list for each nation
There is a lot of repetition here that will be DRY-ed up in the
following commits.

We need to be able to check that the correct content is being rendered
on page in the controller tests, so we need to render the views,
otherwise response.body returns an empty string.