Return the timelines for each nation as one object
This will allow the logic in the view to be DRY-ed up.
The test is very explicit as I think it helps to understand what is
being returned if you can see it in its entirety.

The existing timeline_for_nation method has been made private, so it's
tests have been removed.