Make England the default if no nation selected
This also removes the reliance on a "nation" param existing in the
querystring as we can get the selected nation from the presenter and
also removes some more responsibility from the view.

Passes the selected country directly to the presenter so that it can be
used to set the default nation to England. This means that we can remove
the other checks, as the selected_nation will always have a valid value.