Remove the details element from the nation picker
The original plan was to change the "Change to another nation" text to
instead list the other nations that the user hadn't picked, e.g. "Change
to England, Scotland, or Wales" etc.

However, the nation names in the link would need to be dynamically
updated, which could be confusing without testing, and also looks odd
when the details tab is opened.

There is a concern that leaving the text as it is will mean that user's
won't intuitively understand what is being hidden in the detail. So the
decision was made to remove the details component completely.

We can add back the details component later on when we're more certain
it won't cause users an issue.