Remove JQuery from the Organisation List Filter
We're using an old and unsupported version of jQuery for browser support
reasons.

Rather than upgrade, it's far better to remove our dependence - jQuery
makes writing JavaScript easier, but it doesn't do anything that you
can't do with vanilla JavaScript, because it's all written in
JavaScript.

Once it's removed, we no longer have to worry about upgrading it, and
users don't have to download the jQuery library when they visit GOV.UK.