Hide high_profile_groups heading for organisations without live groups
Organisation page  display High Profile Groups with a status
of `live`, `exempt` or `transitioning`. Other statuses are filtered,
and could leave an empty array.

This adds a check to make sure there is at least one element before rendering
the High Profile Group section. Previously it was possible to display the heading
with no groups.