Order a persons role by seniority
Whitehall has a concept of role seniority [1] which it uses[2] to determine
the order of roles displayed on the government/ministers page [3].
Where 1 is high and 100 is low. We would like to make sure that when collections renders the
organisations/org_name page that the roles under each person are in order
of seniorty.

[1] https://github.com/alphagov/whitehall/blob/main/db/schema.rb#L909
[2] https://github.com/alphagov/whitehall/blob/cf93e6718a45d7aa12e7e0e8dfb196ea041bec07/app/models/minister_sorter.rb#L1
[3] https://www.gov.uk/government/ministers