Apply the big number to the departments page
This adds the big number component to the ministerial departments and organisations page. Included with it, is a change to the js. This is so that we can avoid adding extra classes to the component or having to build a feature therein as this could create some risk of unexpected style override in the future.