Redirect all coronavirus taxons to landing page
Follows on from: https://github.com/alphagov/collections-publisher/pull/1456

We have been given permission to retire the coronavirus hub pages.
The links to the hub pages have already been removed and the content in
the hubs as already been added to the accordions.

Now we need to make sure that the coronavirus child taxons are
redirected to the landing page. If we didn't add this change, we'd still
be redirecting users to the landing page, but there would be one more
jump as they hit the redirect for the hub page first.

Also removes the feature tests that test redirect