Use preferred_url for children on taxon topic pages
We have the concept of an overriding URL which is used when we have a different
page that we'd like to use for a taxonomy landing page.

We need to use this URL on topic pages. There's already a redirect in place,
but it's best to not rely on these if we can as it adds another round-trip
into users' journeys - and therefore, something that can go wrong.

Incidentally, while doing this, I discovered that the test for the sub-taxon
navigation wasn't actually being called by anything. I've added it back
in and fixed it so that it reflects reality.