Refactor `parse_response` method
At least 7 new data points will need to be added to support the new
statistics section design.
In the [spike] the calculations for this data was done in the
`parse_response` method and became very messy. The hope is that
refactoring into smaller methods will make `parse_response` easier to
follow.

[spike]: https://github.com/alphagov/collections/pull/2527/files#diff-54431defd36f31d5ae1e164ecd8eb8f569a17a6c5492c7603fdcb9a45f363d3eR98-R136