Total for last 7 days positive tests
There is no extra data point to collect from coronavirus.data.gov.uk for
this number. Instead we need to make sure have the previous 7 days worth
of cases starting from the first day of data. To make sure this works as
expected, the first value for cases in the tests has been set to nil.

The total for the previous weeks tests is using an `each` rather than a
`map` because `map` adds a `nil` to the array if the condition is false.
Whereas what is required is to completely ignore the value for days
before the first date so we can accurately determine if we do indeed
have 7 days worth of positive test data.

So for example, if the data returned for tests looked like this:

```
[
  {
    "date" => "2021-03-18",
    "newPositiveTests" => nil
  }
  {
    "date" => "2021-03-17",
    "newPositiveTests" => nil
  }
  {
    "date" => "2021-03-16",
    "newPositiveTests" => 5000  <- start counting here
  }
  {
    "date" => "2021-03-15",
    "newPositiveTests" => 4000
  }
  {
    "date" => "2021-03-14",
    "newPositiveTests" => nil   <- need to keep this nil value
  }
  {
    "date" => "2021-03-13",
    "newPositiveTests" => 2000
  }
  .
  .
  .
]
```

We would want to start counting 7 days worth of data from the 16th. So
we can drop the first two days (17th & 18th) where the value for
`newPositiveTests` is `nil`. However we want to keep the `nil` value
from the 14th as this proves that we don't have the full 7 days worth of
data and should not display the section. Without this check we run the
risk of adding data from 8, 9, 10 etc days in the past and calling it
the "last 7 days" data which would be untrue.

`latest_tests` has been updated to return nil if there is no data for
daily or weekly tests so we're never in the position of showing half a
column of data to users. This is merely a precaution, as there are so
many months of COVID data available now that I don't expect this to ever
be a problem.