Allow links in NHS section
This change allows the addition of links in the NHS section. These
links are not H3 elements, but are styled as though they are so as
not to interfere with the overall page design.

This improves the user experience by allowing simple, clean copy to
be created that has a single link to the desired action.

The use of links in headers is not typically done on GOV.UK and may
create some accessibility issues if done. To avoid this situation,
we style the links as though they are H3 elements.