Add i18n-tasks unused keys config
This list of 'ignore_unused' keys tells i18n-tasks which keys to ignore
when it attempts to remove unused keys from the codebase. The keys in
the list are used, but are called upon dynamically (string interpolation
etc), for example: 'schema.#{doc_type}.#{count}'. I18n-tasks default
scanners can't detect these and so we need to add an ignore list.