Use standard metadata for the Brexit page
We don't want custom imagery or code to worry about if we don't need it.

This page is going to be replaced by a standard format soon, and we want
to make that transition ( :-D ) simple by removing as much unnecessary
customisation as possible.

I tweaked a metadata test for og:title because the metadata component sometimes
puts [Withdrawn] in front of the title if the generated test content item
has certain properties.