Remove GOV.UK Account sign-in from /brexit header
GOV.UK account integration has now been removed from the brexit checker,
so there is no longer a need to have a sign in link on the taxon page.

Here we remove the the option that calls this out of slimmer.

We also replace the Header with the new explore banner that we had
previously been overriding to keep our button in place.

And as no other place in collections uses methods from the
personalisation gem, we can also remove that here to keep things tidy.

The sign in link will likely return, but as part of the site-wide
explore banner. So we can likely clear up any GOV.UK Account specific
slimmer stuff after this too.