Use locale files for Organisation Type
The organisation types are specified in the locale files, but we are
instead using hardcoded English text.

This change will mean we will always use the translated name, rather
than always using English.

Additionally, it removes two places that hardcoded lists have to be
edited if an organisation type is added or removed.