Display correct translation for "email"
A change was made to always set the translation for "email" to the
english version for "FOI Contacts" as they come from Whitehall and are
unreliable:
https://github.com/alphagov/collections/pull/1857/commits/b321aaeb285b6da0ffd7fafd21b5902621278a01

However this also had the effect of not translating the text in the
contacts block. This was odd as "telephone" would be translated, but
"email" is not.
See: https://www.gov.uk/government/organisations/land-registry.cy

This change checks if a translation for the page language exists before
using the default English translation. This way we can properly set the
language for the scope of the <p> tag for anyone using a screen reader.