Spike to visualise browse powered by topics
Things to note:

- mainstream_browse_page content items contain a list of all top level browse categories in the attribute
`top_level_browse_pages`. But topic content_items do not contain a list of all the topic categories. I've
added an additional page fetch for the "/topic" content item which contains all the topic categories in
it's "children" attribute but obviously that's just a hack, we should add the list to all topic content items
when they're published.

- mainstram_browse_page content items contain a representation of their parent, children, and siblings.
Topics only contain their parent and child categories. To get this prototype functioning, I've added an addiontal
page fetch. ie when we are on browsetopics/topic/subtopic, the subtopic content item doesn't know who it's siblings are.
So I fetch the parent in order to get the list of siblings. We should add the siblings for subtopics to all topic
content items when they're published.

- There's a concept of links having a curated order, or an alphabetical order. This prototype does not respect the curated
order, because the second_level_ordering attribute is not present on the topic content item.

- Not all topics have a populated details["groups"] attribute.
eg topic/benefits-credits/child-benefit does
but topic/business-enterprise/business-auditing-accounting-reporting

If the groups aren't set, we just show anything tagged to the topic via
https://github.com/alphagov/collections/blob/ac1003904ef6770778ec606e80da14a370e42d7c/app/models/list_set.rb#L43

This is the same behaviour as on /browse at the moment. But just FYI.