Small refactor for template element attribute logic
Having an opening tag inside an if statement and the closing tag outside can be
a bit of a potential gotcha. To avoid this, the logic for the element's
attributes has been separated out to make it more readable and the tag replaced
with Rail's ever useful `content_tag` helper.