Change the branch reference for reusable workflows
This changes the branch reference for reusuable GitHub Action workflow
used to build and push the image. This is workflow is contained in
another repository and to make development easier, swwitching the
reference from the main branch, so a changes to don't need to be merged
to main before testing. This commit should be reverted.