Test Person name in content rather than title
We discovered that the reason that the Rufus Scrimgeour tests were
failing intermittently is due to the page load times.

The test was run 100 times, the tests failed 10% of the time:
```
Finished in 32.02 seconds (files took 6.02 seconds to load)
101 examples, 10 failures
```

When a 10 second sleep was added before the expectation, all of the
tests passed but they took too long to execute. A sleep of 1 or 3
seconds wasn't long enough.

It turns that `have_content` has a built in wait, and waits for the page
to finish loading before testing for the content. Switching to
`have_content ` causes a missing locale error. This will be fixed in the
next commit.

Co-authored-by: Graham Lewis <graham.lewis@digital.cabinet-office.gov.uk>