No longer present special announcements from content item
Special announcements will no longer be pulled in from the content item. This approach is temporary. It is part of work to allow us to remove the govuk-coronavirus-content repo as a dependency. Key name updated as `gettingTestInfo` does exist and is rendered, but the test is testing a non-existent key.

Co-authored-by: Rosa Fox <rosa.fox@digital.cabinet-office.gov.uk>