Add tracking fixes
Correct dimension 1 for level 1 page view, remove a `gem-track-click` from the accordion wrapper so that it's only set on the lists (the 'Show all' button also sets its tracking with a separate attribute). Don't set location or title on the accordion list as this should come from the page view.