Split the second level browse into two templates
The second level browse page is either a list of curated topics displayed in an
accordion component or a single A to Z list.

This was getting too complex to maintain that in a single template, so this has
been split into two templates - one for curated, one for alphabetical - with the
logic to decide which one to use put into the controller. This should make it
easier to iterate on in the future.

Co-authored-by: @hannalaakso.