Add blue bar for users with JavaScript available
When the site-wide coronavirus banner was enabled, there was a bit of
jigger-pokery needed to ensure that the blue bar that's attached to just below
the navigation header was not visible. The coronavirus banner has been
removed - so this isn't needed any longer.

However, it does raise an issue that the blue bar isn't implemented consistently
across different applications - for example, the new topics pages have a full
width dark header which require a constrained width blue bar; a taxon has a full
width dark header that doesn't require a blue bar at all; the coronavirus page
has full width yellow bar (the colour is different, but the placement and
dimensions are the same as the blue coloured bar). Mentioning this problem here
as it's worth noting but out of scope for this task.