Do not add locale files for coronavirus pages
The I18n-tasks gem which is used by Rails translation manager
under the hood, is failing to parse the coronavirus show page.
This is preventing the rake:add_mising task from completing.

There is currently no locale file for the coronavirus pages,
all the content is hard coded in the view. Adding the strings
to a new locale file does not fix the problem.

In the interest of getting the rest of the locale files in sync
I've exlcuded the coronavirus pages for now.