Update i18n-tasks to ignore a key
When running bundle exec rake:remove_unused, rails translation
manager is removing the key organisations.works_with_statement as it
isn't present in any of the view files. That's because it's being
dynamically set here:
https://github.com/alphagov/collections/blob/main/app/presenters/organisations/index_presenter.rb#L66-L68

As per the docs we should update the yaml config file so that it
ignores this key.