Render document lists on Topical Event page
This renders the publications, consultations and announcements sections.

Whilst we've tried to keep this as visually similar to the existing page that's rendered by whitehall as possible there are slight visual differences in the name of simplicity. Whitehall has a somewhat complex and snowflakey way of rendering these lists. We've simply used the existing solution from publishing components which makes the view consistent across other parts of GOV.UK. The other ramification of this change is the apparent removal of tracking on links within these document lists. However, we've had a word with a performance analyst and it appears this tracking hasn't worked in some time and no one seems to care. We're therefore making the decision to not add traking and remove the complexity that whitehall introduced for good. If in the future this needs to be revisited then changes can be made to the publishing components gem to add optional GA link tracking on document lists.