Add Afghanistan foreign travel advice case
The uk response in Afghanistan is a unique Topical event. I think as some 2nd line work a case was added whereby that Topical event would have a link to the travel advice for Afghanistan (commit here https://github.com/alphagov/whitehall/commit/40fc29efa2f733c51c615592d7556c77e3a58fac). This ports that work over and adds tests. It's not ideal to have this hardcoded and there's work in the pipeline to allow travel advice to be tagged to topical events (trello - https://trello.com/c/hmcqxhAr/579-allow-travel-advice-to-be-tagged-to-topical-events).