Add dimension 114 to click event on show/hide all
When clicking Show/Hide All sections on curated browse pages the index
of this section in the accordion was not being passed in the click
event which made analysis to determine how the order of sections
impacts click rate more difficult. This index (of 0) is now passed
when a user clicks 'Show/Hide All' after an update to the component
now allows the passing of track options to the show/hide all button.