Add phase to fix failing test
This test was failing as there was no value for `phase` defined. However
because we used `better_errors` and the expectation is not to find
certain elements, the test always passed despite the response being a
500 error.

Now `better_errors` is not being used for tests, we see the error and
have fixed it accordingly.