Delete old Miller column style templates
We ran an A/B test to compare the current "miller column" style templates (A) against the new more accessible and mobile friendly GOV.UK topics templates (B) . The results of the test have given us confidence that we can retire the old templates and make the new templates the default ones (with some further iterations,  see next commits).

Delete the templates and the controller logic for the old templates.