Truncate description for featured documents when they exceed limit
In Whitehall, the description is truncated at 160 characters (or less to avoid truncating in the middle of a word)  when displaying featured documents. Publishing apps display a warning when creating documents with summaries over 160 characters, but does not enforce this. This means that it is down to frontend apps to truncate overly long content.