Fix padding issue for organisations list
The default padding for unordered lists is 40px, as set by user-agent-stylesheets. We want to override this to render organisation logos sensibly. This is an issue specifically with topical events pages which currently have hard-coded logic in place to deal with the custom requirement of displaying non-government organisation logos (example page: /topical-events/first-world-war-centenary).

This change ensures that these logos are rendered in line with the government organisation logos.