Divide the topic 2 lists into sections
As part of iterating the topics level 2 template we want to break the curated lists with their headings into sections divided by borders. This might help users to scan the potentially quite long topic pages since we're no longer using the accordions to organise the content.

Add top border to all the curated list sections. Add a bottom border to the last section when there are no related topics below it.