Pass requests for subtopics of interest to browse controllers
Collections handles the rendering of browse pages and topic pages.
If we want to render a level two topic page as a second level browse page,
we need to add a routing constraint: Check the request path against config
file, and if it matches, forward the request to the second level browse
controller instead of to the topic controller.