Show subtopics in the browse style
Some of the attributes included on the mainstream browse schema are
relied on to build the miller columns. Once the NewBrowse AB test has
completed, and we are showing all users a version of the New Browse
templates, we can remove these accessor methods from the Topic model,
and simplify the Mainstream Browse model as well.