Fix feature tests for Browse pages
We no longer have the Miller Columns and the feature specs need updating.

Remove dependencies on gem component classes from the browse feature tests. Test that the correct content is used instead. Remove functionality that tests whether the cards component outputs a list of items alphabetically, this should be tested at the gem level if needed.