Add accordion & list templates for curated level 2
We will test the existing B variant with accordions alongside the C variant
which uses lists. Add templates for both.

Remove the accordion from the curated page templates, to be replaced with a
call to the relevant partial in the next commit.

For the curated list, copy over the styles of the topics A-Z list, to be
further iterated in the next commits.