Switch between B and C variants
Add logic to switch between B and C variants. The variants use the same templates, the only change is how the content items on the Level 2 curated pages are presented (accordions or lists).

Switching the variant based on the query string is only for testing purposes, we will implement the proper logic in a later PR.