Rename new topic pages to browse
Sometime during the building of the first iteration of the new browse templates we took the implicit decision to refer to them as `topic-page` in the scss and some of the markup. This was probably to differentiate them from the miller columns scss and markup which used `browse` in scss and markup.

Now that we've removed the miller columns, rename the new browse page styles, markup and file names to use `browse` consistently. This will avoid confusion between having references to both `browse` and `topic-page` in the codebase, especially as there are also separate styles and templates under `app/views/topics`.