Switch between A and B variants
Add logic to switch between A and B variants. The variants use the same templates, the only change is how the content items on the Level 2 curated pages are presented (accordions or lists). Make the curated lists the A (default) template.

Switching the variant based on the query string is only for testing purposes, we will implement the proper logic in a later PR.