Wrap the browse level 2 content into sections
Break the curated lists with their headings into browse sections divided by borders. This might help users to scan the potentially quite long browse pages since this variant of the curated browse level 2 pages no longer uses accordions to organise its content.

Wrap also the A-Z list and the accordion in these browse sections to standardise the top and bottom margins of the browse content at level 2. This is useful since:
- the accordion does not accept custom bottom margins and the Design System responsive classes don't offer enough spacing on mobile
- the existing A-Z list needs to be wrapped in a div to stop its bottom margin collapsing with the 'More on this topic' top margin below it (the 'More on this topic' isn't always present and therefore sets a top margin for its spacing).

Add top border to all sections that are not the first, or only, section on the pages. This adds a border to all the curated list sections (except the first one) but not to the accordion or the A-Z list.

Note: This approach can further be tidied up and simplified when the AB test has completed and we remove either the curated accordion or curated list partial.