Deprecate old Miller column style templates
We ran an A/B test to compare the current "miller column" style templates (A) against the new more accessible and mobile friendly GOV.UK topics templates (B) . The results of the test have given us confidence that we can retire the old templates and make the new templates the default ones (with some further iterations,  see next commits).

However, we need to deprecate the Miller column templates until some time after the new templates have been deployed. This is because if a user is on the Miller column pages when the new templates get deployed, we need to still be able to call the endpoints for the JSON blobs whilst they are viewing the page.

Rename the Miller column templates to 'old-' to avoid file name clashes with the new ones. We'll add the switching logic for the new templates in a later commit.