Add bypass for 404 item not found error
Usually we throw a `404 Not Found` if we don't have a document in the content store.
This is a naive bypass for now so we can deliver with hardcoded values.

Also does a minor change to call to content_item.yml rather than content.yml

It will be removed when we publish the special route
(https://docs.publishing.service.gov.uk/manual/publish-special-routes.html)