Add the yml for content for Cost of living Hub
As there is no content in the content store yet for the Cost of living
Hub we are using a YML file as a stand-in. This contains all the content and the
metadata for the page.

As the title for page has changed, the test has been updated so that it
can pass.