Remove hardcoded list of document types
We had previously hardcoded a list of document types to retrieve when
getting announcements to display on topical event pages.

These can be replaced with the data in the `govuk_document_types` gem.

This means the list won't get out of date if (for example) new content
types are added or removed from the group.