Add document sections to world location news page
This page is currently rendered by Whitehall. This is part of the work
to bring the rendering into Collections.

We have tried to reuse as much code as possible from topical events in
rendering this section.

The horizontal line between each section has been updated to match the
style of Topical Event pages. This differs from the rendering in
Whitehall.