Show guidance and regulation rather than detailed guides on topical events
We have switched to using the content_purpose_supergroup of guidance_and_regulation rather than the filter_format of detailed_guidance for topical events. This aligns more with other content types in collections, e.g. organisations, which prefer content_purpose_supergroups instead of specific document types. This also means that the link to search is now more relevant, as it links to all guidance and regulation for the topical event, rather than just to detailed guides.