update topical events announcements to use content purpose supergroup
Previously, we were specifying the individual document types that made up the supergroup of announcements. Instead, we can simplify by switching to the content purpose supergroup of news and communications and avoid essentially replicating a hard-coded list that is already found in govuk-document-types.

We are excluding some document types via the reject_content_purpose_subgroup filter - this should match what is done for other pages rendered by collections.