Change default cache time to 5 mins
This is a very quick fix to reduce the global caching time of this application from 30 mins to 5 mins. This is a temporary fix to remedy RFC 144 not being implemented fully in Collections: https://github.com/alphagov/collections/pull/2659

The need for a temporary fix is because we are doing a reshuffle today, and want to avoid people manually clearing cache entries.

There shouldn't be any need for pages in this application to be caching pages for longer than 5 mins.