[WIP] Order announcements by popularity and by updated in last 7 days
Docs: https://github.com/alphagov/search-api/blob/2675a59/docs/public-api/faceted-search.md#filtering-parameters

For date fields, values are date ranges. These are specified as comma separated
lists of key:value parameters, where key is one of from or to, and the value is
an ISO formatted date (with no timezone). UTC is assumed for all dates. Date
ranges are inclusive of their endpoints.

For example: from:2014-04-01 00:00,to:2014-04-02 00:00 is a range for 24 hours
from midnight at the start of April the 1st 2014, including midnight that day
or the following day.