Use request path for World Location News
We are currently using a path that doesn't contain the locale when
requesting World Location News from the content store.

This means we can't currently show anything other than English content.

Changing to `request.path` means we will include the locale when
retrieving the content.

This is consistent with the usage on organisation pages.