Normalize locale files
This normalizes the locale files in accordance with the rules defined in
`rails_translation_manager`.

The output of this commit is from the rake task `rake
translation:normalize`.

If we don't normalize the files now, then we'll get a git diff that
changes irrelevant lines when we import keys in a later commit.