Switch back to using the featured document doc type from the content item
This is already translated so we don't need to translate here, we can just directly render the document type.

Additionally, this fixes a bug whereby some non-english translations could not render. Parameterize is for urls, and strips out any characters that cannot appear in a url. This means that, for languages such as arabic, parameterize returns an empty string.

This was causing a bug for any world location news pages with featured documents where the document type (which is translated) is entirely constituted of characters that cannot appear in urls (for exmaple, in languages which don't use latin scripts). Before this change, we were getting the document type to display with the following code:

`shared.schema_name.#{document['document_type']&.parameterize(separator: '_')}`

where the 'document' was the content item we are rendering.

If this was working, we'd expect the code to be looking for something like the following in locale files:

'shared.schema_name.speech'

However, when the document type in the content item contains only non-url allowable characters, we were looking for 'one' of the following string in the locale files:

'shared.schema_name.'

so the string after the final dot is empty. This causes the page to error upon rendering with the following error:

'I18n::InvalidPluralizationData'

Additionally, since we were already translating the document type in the content item, even in cases where this was working for translated pages, it was only working because a translation was not found (since the key that we were looking up contained the translated document type) and therefore we defaulted back to just rendering the document type from the content item.